postgres-sync/requirements.txt
firebase-admin==6.4.0
psycopg2-binary==2.9.9
